refactor: moving the top level methods
